requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.4.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
coverage==7.11.0
dj-database-url==3.0.1
Django==5.2.2
django-environ==0.12.0
django-phonenumber-field==8.3.0
factory-boy==3.3.0
Faker==37.12.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
iniconfig==2.3.0
joblib==1.5.2
multidict==6.7.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
phonenumbers==9.0.16
pillow==11.3.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-mock==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.3
responses==0.24.1
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
twilio==9.8.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
whitenoise==6.11.0
yarl==1.22.0
